Regenerating bindings for 10.0.22000.0

+++ sources/Interop/Windows/Windows/um/winnt/POWER_SESSION_RIT_STATE.cs
@@ -1,6 +1,6 @@
 // Copyright © Tanner Gooding and Contributors. Licensed under the MIT License (MIT). See License.md in the repository root for more information.
 
-// Ported from um/winnt.h in the Windows SDK for Windows 10.0.20348.0
+// Ported from um/winnt.h in the Windows SDK for Windows 10.0.22000.0
 // Original source is Copyright © Microsoft. All rights reserved.
 
 namespace TerraFX.Interop.Windows;
@@ -13,6 +13,6 @@ public partial struct POWER_SESSION_RIT_STATE
     public byte Active;
 
     /// <include file='POWER_SESSION_RIT_STATE.xml' path='doc/member[@name="POWER_SESSION_RIT_STATE.LastInputTime"]/*' />
-    [NativeTypeName("DWORD")]
-    public uint LastInputTime;
+    [NativeTypeName("DWORD64")]
+    public ulong LastInputTime;
 }
